US32 - Refactor Dateutility
