+++ TruckFeed/TruckFeed/EndDateSwitchCell.swift
@@ -16,25 +16,14 @@ class EndDateSwitchCell: UITableViewCell {
     public var delegate: AddScheduleDataProvider?
     
     override func awakeFromNib() {
-        endDateLabel.text = getCurrentDateTime()
-        endDateSwitch.addTarget(self, action: #selector(endTimeSwitchToggled), for: UIControlEvents.valueChaavgvanged)
+        endDateLabel.text = DateUtility.getCurrentDateTime()
+        endDateSwitch.addTarget(self, action: #selector(endTimeSwitchToggled), for: UIControlEvents.valueChanged)
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: "com.lovelejess.endDateLabelSelected"), object: nil, queue: nil, using: updateEndDateLabel)
     }
     
     func endTimeSwitchToggled(){
         delegate!.reloadTableData()
     }
-    
-    public func getCurrentDateTime() -> String {
-        let currentDateTime = Date()
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "MM/dd/yy hh:mm a"
-        dateFormatter.amSymbol = "AM"
-        dateFormatter.pmSymbol = "PM"
-        let date = dateFormatter.string(from: currentDateTime)
-        return date
-    }
-
     
     private func updateEndDateLabel(notification: Notification) -> Void {
         if let userInfo = notification.userInfo {
