+++ TruckFeed/TruckFeed/StartDateSwitchCell.swift
@@ -16,23 +16,13 @@ class StartDateSwitchCell: UITableViewCell {
     public var delegate: AddScheduleDataProvider?
     
     override func awakeFromNib() {
-        startDateLabel.text = getCurrentDateTime()
+        startDateLabel.text = DateUtility.getCurrentDateTime()
         startDateSwitch.addTarget(self, action: #selector(startTimeSwitchToggled), for: UIControlEvents.valueChanged)
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: "com.lovelejess.startDateLabelSelected"), object: nil, queue: nil, using: updateStartDateLabel)
     }
     
     func startTimeSwitchToggled(){
         delegate!.reloadTableData()
-    }
-    
-    public func getCurrentDateTime() -> String {
-        let currentDateTime = Date()
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "MM/dd/yy hh:mm a"
-        dateFormatter.amSymbol = "AM"
-        dateFormatter.pmSymbol = "PM"
-        let date = dateFormatter.string(from: currentDateTime)
-        return date
     }
     
     private func updateStartDateLabel(notification: Notification) -> Void {
